requirements.txt
Flask
opencensus-ext-azure
opencensus-ext-flask